django/requirements.txt
psycopg2
coverage==4.0.3
cssselect==0.9.1
Django==1.9.6
django-appconf==1.0.2
django-compressor==2.0
django-registration==2.0.4
djangohelpers==0.18.1
factory-boy==2.6.0
fake-factory==0.5.3
flake8==2.5.1
funcsigs==1.0.2
lxml==3.6.0
mccabe==0.3.1
mock==1.3.0
pbr==1.9.1
pep8==1.7.0
pyflakes==1.0.0
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
